requirements.txt
Django>=5.2.6
djangorestframework>=3.14
djangorestframework-simplejwt
psycopg2-binary>=2.9
drf-spectacular
pillow
celery~=5.5.3
pytest~=8.4.2